Update README, delete TypeNameStyle enum, and enhance build properties
This commit removes the `TypeNameStyle` enum as it is no longer necessary. Comprehensive alterations have been made to the `README.md` file to improve clarity and provide detailed instructions for installation and usage. Additionally, more details were added to the `Directory.Build.props` to increase the project's overall configurability and visibility.