Add new library for dumping types as initialization code
This update introduces a new library called "Frank.Reflection.Dump". This library allows developers to dump the contents of a type into initialization code. This can be useful for debugging, logging, and generating unit test cases. Additionally, Dependency updates, a few changes to the CSPROJ files related to packaging, and minor adjustments to some of the existing files, including in-line comments and changes to string split methods, have been included.